Write Test Suite for 'The menu'
Write a new test suite named "The menu".
Write a test that ensures the menu element is hidden by default. You'll
have to analyze the HTML and the CSS to determine how we're performing
the hiding/showing of the menu element.
Write a test that ensures the menu changes visibility when the menu
icon is clicked. This test should have two expectations: does the menu
display when clicked and does it hide when clicked again.